Authorization with policies

+++ app/Http/Controllers/Api/EventController.php
@@ -25,7 +25,7 @@ class EventController extends Controller
     public function index()
     {
         // $this -> shouldIncludeRelation('user');
-
+        Gate::authorize('viewAny', Event::class);
         $query = $this->loadRelationships(Event::query());
 
         return EventResource::collection(
@@ -43,7 +43,7 @@ class EventController extends Controller
     {
 
 
-
+        Gate::authorize('create', Event::class);
         $event = Event::create([
             ...$request->validate([
                 'name' => 'required|string|max:255',
@@ -63,7 +63,7 @@ class EventController extends Controller
      */
     public function show(Event $event)
     {
-
+        Gate::authorize('view', $event);
         return new EventResource( $this->loadRelationships($event));
     }
 
@@ -96,7 +96,7 @@ class EventController extends Controller
      */
     public function destroy(Event $event)
     {
-
+        Gate::authorize('delete', $event);
         $event->delete();
 
         return response( status: 204);
